backend/requirements.txt
fastapi
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
matplotlib
matplotlib-inline
mlflow
numpy
pandas
pillow
pyarrow
python-dotenv
requests
scikit-learn
seaborn 